[Bridge] organized Twig/Translation bridge classes
