+++ tests/Symfony/Tests/Bridge/Twig/Extension/FormExtensionTableLayoutTest.php
@@ -16,7 +16,7 @@ require_once __DIR__.'/Fixtures/StubFilesystemLoader.php';
 
 use Symfony\Component\Form\FormView;
 use Symfony\Bridge\Twig\Form\FormExtension;
-use Symfony\Bridge\Twig\Extension\TranslationExtension;
+use Symfony\Bridge\Twig\Translation\TranslationExtension;
 use Symfony\Tests\Component\Form\AbstractTableLayoutTest;
 use Symfony\Tests\Bridge\Twig\Extension\Fixtures\StubTranslator;
 use Symfony\Tests\Bridge\Twig\Extension\Fixtures\StubFilesystemLoader;
